scanner/requirements.txt
Deprecated==1.2.13
mysql-connector-python==8.0.28
packaging==21.3
protobuf==3.19.4
pyparsing==3.0.7
redis==4.1.1
wrapt==1.14.0
sqlalchemy==1.4.32
mysqlclient==2.1.0